+++ familyhelper-assets-impl/src/main/java/com/dwarfeng/familyhelper/assets/impl/util/FtpConstants.java
@@ -8,8 +8,15 @@ package com.dwarfeng.familyhelper.assets.impl.util;
  */
 public final class FtpConstants {
 
-    public static final String PATH_ITEM_COVER = "familyhelper-assets/item-cover";
-    public static final String PATH_ITEM_FILE = "familyhelper-assets/item-file";
+    /**
+     * @since 1.0.8
+     */
+    public static final String[] PATH_ITEM_COVER = new String[]{"familyhelper-assets", "item-cover"};
+
+    /**
+     * @since 1.0.8
+     */
+    public static final String[] PATH_ITEM_FILE = new String[]{"familyhelper-assets", "item-file"};
 
     private FtpConstants() {
         throw new IllegalStateException("禁止实例化");